requirements.txt
requests
pandas
python-dotenv
pytest
pytest-cov
pytest-mock